requirements.txt
httpx>=0.24.0
beautifulsoup4>=4.12.0
feedparser>=6.0.10
aiohttp>=3.8.0
lxml>=4.9.0
requests>=2.31.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0